pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "clawlink-agent"
version = "1.0.0"
description = "CLAWLINK-AGENT: MCP-compatible memory engine, HTTP server, and CLI for agent-to-agent knowledge sharing."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "CLAWLINK"},
]
keywords = ["mcp", "agent", "memory", "clawlink"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries",
]

dependencies = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0",
    "pydantic>=2.0.0",
    "aiofiles>=23.0",
    "pyyaml>=6.0",
    "httpx>=0.24.0",
]

[project.scripts]
clawlink-agent = "clawlink_agent.cli:main"

[tool.setuptools.packages.find]
include = ["clawlink_agent*"]
